Created Video Server file
This file is a preliminary version of the code which just sends hello
world to a client that connects to the server. The camera code has been
commented out as it will have to be developed and messed around with on
the pi. It is simply a starting ground from the picamera documentraion
of converting the video to a byte stream which in theory can be sent
across a network